feat: update Purchase and PurchaseLineItem to use BigDecimal for quan…

+++ src/main/java/com/spiceflow/backend/purchase/dto/response/PurchaseLineItemResponse.java
@@ -12,8 +12,8 @@ public record PurchaseLineItemResponse(
     String productName,
     String productSku,
     
-    Integer noOfBoxes,
-    Integer soldQuantity,
+    BigDecimal noOfBoxes,
+    BigDecimal soldQuantity,
     String unitType,
     BigDecimal rate,
     BigDecimal amount
